[fix] Enforce compatible Node.js and NPM version
Currently, the Ethical Framework uses the async/await feature which
requires Node.js 7.6 or above. It also relies on a flat tree dependency
to resolve modules, which is present only on NPM 3 and above. One
solution is to specify the "engines" attribute in the package.json to
have NPM warn the user at the time of install.